#3 feat: implementado as rotas iniciais e feita a estrutura do projeto.
